include: Fix integer type of dh_pn and dh_pmax
Fixes T188

+++ include/libssh/crypto.h
@@ -105,7 +105,7 @@ struct ssh_crypto_struct {
     bignum shared_secret;
     struct dh_ctx *dh_ctx;
 #ifdef WITH_GEX
-    size_t dh_pmin; int dh_pn; int dh_pmax; /* preferred group parameters */
+    size_t dh_pmin; size_t dh_pn; size_t dh_pmax; /* preferred group parameters */
 #endif /* WITH_GEX */
 #ifdef HAVE_ECDH
 #ifdef HAVE_OPENSSL_ECC
